# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.BookHeadingEdit:
    properties:
      book_id:
        example: 797
        type: integer
      content:
        type: string
      heading_id:
        example: 10
        type: integer
      published_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      status:
        example: draft
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      updated_by:
        type: string
    type: object
  entity.BookMetadataEdit:
    properties:
      bibliography:
        type: string
      book_id:
        example: 797
        type: integer
      category_id:
        type: integer
      cover_url:
        type: string
      description:
        type: string
      display_title:
        type: string
      hint:
        type: string
      notes:
        type: string
      published_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      status:
        example: draft
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      updated_by:
        type: string
    type: object
...
  entity.BookPageEdit:
    properties:
      book_id:
        example: 797
        type: integer
      content_html:
        type: string
      content_text:
        type: string
      page_id:
        example: 1
        type: integer
      published_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      status:
        example: draft
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      updated_by:
        type: string
    type: object
...
  entity.BookSourceEditRevision:
    properties:
      actor_id:
        type: string
      asset_type:
        example: page
        type: string
      book_id:
        example: 797
        type: integer
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      heading_id:
        example: 10
        type: integer
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      origin:
        example: rest
        type: string
      page_id:
        example: 1
        type: integer
      snapshot:
        type: object
      version:
        example: 3
        type: integer
    type: object
...
  entity.Translation:
    properties:
      destination:
        example: en
        type: string
      original:
        example: текст для перевода
        type: string
      source:
        example: auto
        type: string
      translation:
        example: text for translation
        type: string
    type: object
...
  entity.TranslationHistory:
    properties:
      history:
        items:
          $ref: '#/definitions/entity.Translation'
        type: array
    type: object
...
  v1.Error:
    properties:
      code:
        example: invalid_request_body
        type: string
      details: {}
      error:
        example: message
        type: string
      message:
        example: message
        type: string
      request_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      retry_after:
        example: 60
        type: integer
    type: object
...
  v1.SourceEditRevisionList:
    properties:
      revisions:
        items:
          $ref: '#/definitions/entity.BookSourceEditRevision'
        type: array
      total:
        example: 42
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Surau classical book reader API
  title: Go Clean Template API
  version: "1.0"
paths:
...
  /editorial/books/{book_id}/headings/{heading_id}/draft:
    get:
      description: Get the current source kitab heading title draft. Requires editor
        or admin role.
      operationId: editorial-get-source-heading-draft
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: integer
      - description: Heading ID
        in: path
        name: heading_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.BookHeadingEdit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Get source heading draft
      tags:
      - editorial
  /editorial/books/{book_id}/metadata-draft:
    get:
      description: Get the current source kitab metadata draft. Requires editor or
        admin role.
      operationId: editorial-get-source-metadata-draft
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.BookMetadataEdit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Get source metadata draft
      tags:
      - editorial
  /editorial/books/{book_id}/pages/{page_id}/draft-revisions:
    get:
      description: List the newest-first revision history for one page draft. Requires
        editor or admin role.
      operationId: editorial-list-page-draft-revisions
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: integer
      - description: Page ID
        in: path
        name: page_id
        required: true
        type: integer
      - description: Page size (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SourceEditRevisionList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: List page draft revisions
      tags:
      - editorial
  /editorial/books/{book_id}/pages/{page_id}/draft-revisions/{revision_id}/restore:
    post:
      description: Replay a previous page draft snapshot as the current draft; the
        restore creates a new revision. Requires editor or admin role.
      operationId: editorial-restore-page-draft-revision
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: integer
      - description: Page ID
        in: path
        name: page_id
        required: true
        type: integer
      - description: Revision ID
        in: path
        name: revision_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.BookPageEdit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Restore page draft revision
      tags:
      - editorial
...
  /translation/history:
    get:
      description: Show all translation history for current user
      operationId: history
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.TranslationHistory'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Show history
      tags:
      - translation
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"